Enhances UI and adds favorite toggle animation
Adds a subtle shadow to the app's logo for visual appeal.

Uses ultra thin material for the background of the pokemon card.

Adds a smooth animation when toggling the favorite state of a Pokemon.

Increases the scale of the Pokemon image in the detail view
for better visual presentation.

Improved dark mode contrast.